requirements.txt
# --- Bases de Datos ---
pymongo>=4.8.0
cassandra-driver>=3.29.2
pydgraph>=21.3.2

# --- Análisis y Utilidades ---
pandas>=2.2.2
numpy>=1.26.4
requests>=2.32.3
tabulate

# --- Fixes para Mac M1 / Apple Silicon ---
# Forzamos versiones modernas de gRPC y Protobuf para evitar lentitud
grpcio>=1.60.0
protobuf>=4.21.0